+++ zarrs_metadata/src/v2/array/codec/vlen_utf8.rs
@@ -1,0 +1,4 @@
+/// The identifier for the `vlen-utf8` codec.
+pub const IDENTIFIER: &str = "vlen-utf8";
+
+pub use crate::v3::array::codec::vlen_v2::VlenV2CodecConfigurationV1;